web/lib/firebase/api-call -> common/api, web/lib/firebase/api
