+++ web/components/bet-panel.tsx
@@ -19,8 +19,8 @@ import { getBinaryCpmmBetInfo } from 'common/new-bet'
 import { Title } from './title'
 import { User } from 'web/lib/firebase/users'
 import { Bet, LimitBet } from 'common/bet'
-import { APIError, placeBet } from 'web/lib/firebase/api-call'
-import { sellShares } from 'web/lib/firebase/api-call'
+import { APIError, placeBet } from 'web/lib/firebase/api'
+import { sellShares } from 'web/lib/firebase/api'
 import { AmountInput, BuyAmountInput } from './amount-input'
 import { InfoTooltip } from './info-tooltip'
 import { BinaryOutcomeLabel } from './outcome-label'
