+++ web/components/contract/quick-bet.tsx
@@ -22,14 +22,14 @@ import {
 import { useState } from 'react'
 import toast from 'react-hot-toast'
 import { useUserContractBets } from 'web/hooks/use-user-bets'
-import { placeBet } from 'web/lib/firebase/api-call'
+import { placeBet } from 'web/lib/firebase/api'
 import { getBinaryProb, getBinaryProbPercent } from 'web/lib/firebase/contracts'
 import TriangleDownFillIcon from 'web/lib/icons/triangle-down-fill-icon'
 import TriangleFillIcon from 'web/lib/icons/triangle-fill-icon'
 import { Col } from '../layout/col'
 import { OUTCOME_TO_COLOR } from '../outcome-label'
 import { useSaveBinaryShares } from '../use-save-binary-shares'
-import { sellShares } from 'web/lib/firebase/api-call'
+import { sellShares } from 'web/lib/firebase/api'
 import { calculateCpmmSale, getCpmmProbability } from 'common/calculate-cpmm'
 import { track } from 'web/lib/service/analytics'
 import { formatNumericProbability } from 'common/pseudo-numeric'
