+++ web/components/groups/create-group-button.tsx
@@ -9,7 +9,7 @@ import { Title } from '../title'
 import { FilterSelectUsers } from 'web/components/filter-select-users'
 import { User } from 'common/user'
 import { MAX_GROUP_NAME_LENGTH } from 'common/group'
-import { createGroup } from 'web/lib/firebase/api-call'
+import { createGroup } from 'web/lib/firebase/api'
 
 export function CreateGroupButton(props: {
   user: User
