+++ web/components/limit-bets.tsx
@@ -5,7 +5,7 @@ import { getFormattedMappedValue } from 'common/pseudo-numeric'
 import { formatMoney, formatPercent } from 'common/util/format'
 import { sortBy } from 'lodash'
 import { useState } from 'react'
-import { cancelBet } from 'web/lib/firebase/api-call'
+import { cancelBet } from 'web/lib/firebase/api'
 import { Col } from './layout/col'
 import { LoadingIndicator } from './loading-indicator'
 import { BinaryOutcomeLabel, PseudoNumericOutcomeLabel } from './outcome-label'
