+++ web/components/liquidity-panel.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react'
 import { CPMMContract } from 'common/contract'
 import { formatMoney } from 'common/util/format'
 import { useUser } from 'web/hooks/use-user'
-import { addLiquidity, withdrawLiquidity } from 'web/lib/firebase/api-call'
+import { addLiquidity, withdrawLiquidity } from 'web/lib/firebase/api'
 import { AmountInput } from './amount-input'
 import { Row } from './layout/row'
 import { useUserLiquidity } from 'web/hooks/use-liquidity'
