+++ web/components/notifications-icon.tsx
@@ -6,7 +6,7 @@ import { usePrivateUser, useUser } from 'web/hooks/use-user'
 import { useRouter } from 'next/router'
 import { useUnseenPreferredNotificationGroups } from 'web/hooks/use-notifications'
 import { NOTIFICATIONS_PER_PAGE } from 'web/pages/notifications'
-import { requestBonuses } from 'web/lib/firebase/api-call'
+import { requestBonuses } from 'web/lib/firebase/api'
 import { PrivateUser } from 'common/user'
 
 export default function NotificationsIcon(props: { className?: string }) {
