+++ web/pages/create.tsx
@@ -6,7 +6,7 @@ import Textarea from 'react-expanding-textarea'
 import { Spacer } from 'web/components/layout/spacer'
 import { useUser } from 'web/hooks/use-user'
 import { Contract, contractPath } from 'web/lib/firebase/contracts'
-import { createMarket } from 'web/lib/firebase/api-call'
+import { createMarket } from 'web/lib/firebase/api'
 import { FIXED_ANTE } from 'common/antes'
 import { InfoTooltip } from 'web/components/info-tooltip'
 import { Page } from 'web/components/page'
