+++ web/pages/link/[slug].tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router'
 import { useState } from 'react'
 import { SEO } from 'web/components/SEO'
 import { Title } from 'web/components/title'
-import { claimManalink } from 'web/lib/firebase/api-call'
+import { claimManalink } from 'web/lib/firebase/api'
 import { useManalink } from 'web/lib/firebase/manalinks'
 import { ManalinkCard } from 'web/components/manalink-card'
 import { useUser } from 'web/hooks/use-user'
